requirements.txt
#
# This file is autogenerated by pip-compile
# To update, run:
#
#    pip-compile --output-file requirements.txt requirements.in
#
progressbar2==3.18.1
python-utils==2.1.0       # via progressbar2
six==1.10.0               # via python-utils
